refactores to clean up before janus
