--- public/js/chatBox.js
@@ -1,5 +1,0 @@
-var messageContainer = document.getElementById('message-container');
-var sendButton = document.getElementById('send-button');
-sendButton.addEventListener('click', function () {
-  messageContainer.scrollTop = messageContainer.scrollHeight;
-});
