+++ public/stats.html
@@ -137,7 +137,7 @@
         };
     }
   </script>
-  <script src="/stats.js" type="text/javascript"></script>
+  <script src="/stats.js" type="module"></script>
 </body>
 
 </html>
